change:[DAG] change double type to integer

+++ database/migrations/2024_06_26_022054_create_barangs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama');
             $table->string('kategori');
-            $table->double('harga');
+            $table->integer('harga');
             $table->timestamps();
         });
     }
